Remove deprecated methods
